PORTAL-360: Implement - Responsive Footer (#166)
* Deleted useless tag

Deleted useless tag

* Added line-break: anywhere; for contact

Added line-break: anywhere; for contact

* fixed Scroll button

fixed Scroll button

* Change back to top img to arrow up icon

Change back to top img to arrow up icon

* Added responsive to footer tablet and mobile

Added responsive to footer tablet and mobile

* commented out outline on navbar

commented out outline on navbar

* cleared up the footer data

cleared up the footer data

* Cleaned up: delete useless header footer and navbar

Cleaned up: delete useless header footer and navbar

* Changed headerData structure

Changed headerData structure

* put navbarData into globalHeaderData

put navbarData into globalHeaderData
