--- src/components/NavBar/NavBarView.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import { NavBar } from 'bento-components';
-import {
-    navBarData, navBarCartData, navBarstyling
-  } from '../../bento/navigationBarData';
-
-const BentoNavBar = () => {
-  return (
-    <>
-      <NavBar
-        navBarData={navBarData}
-        navBarCartData={navBarCartData}
-        navBarstyling={navBarstyling}
-      />
-    </>
-  )
-};
-
-export default BentoNavBar;
