--- src/components/ScrollButton/ScrollButtonStyles.css
@@ -1,52 +1,0 @@
-#stt {
-    background: #007bbd;
-    border-top-left-radius: 100%;
-    color: #fff;
-    position: fixed;
-    right: 0;
-    bottom: 0;
-    height: 80px;
-    width: 80px;
-    font-family: 'Open Sans';
-    font-style: normal;
-    font-weight: 700;
-    font-size: 12.16px;
-    line-height: 16px;
-    text-align: center;
-    padding: 36px 4px 0 18px;
-    text-decoration: none;
-    line-height: 1.2;
-    transition: all 0.25s ease-out;
-    z-index:10;
-}
-
-#stt:active {
-    outline: solid 4px #2491ff;
-}
-
-#stt::after {
-    content: "";
-    display: none;
-}
-
-/*
-Tablet/Phone less than 768 px width and greater than greater than 375
-*/
-@media (max-width: 907px){
-    #stt-span {
-        display: none;
-    }
-
-    #stt::after {
-        background:none;
-        background-color: #fff;
-        -webkit-mask: url("./Up\ Arrow.svg") no-repeat center/contain;
-        mask: url("./Up\ Arrow.svg") no-repeat center/contain;
-        display: inline-block;
-        height: 4ex;
-        margin-left: 1px;
-        vertical-align: middle;
-        width: 4ex;
-        color:white;
-    }
-}
